+++ extensions/typora-note-creator/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Typora Note Creator Changelog
 
-## [Initial Version] - 2025-08-22
+## [Initial Version] - {PR_MERGE_DATE}
+
+## [Bug Fixes] - {PR_MERGE_DATE}
+
+- Handle undefined notesDir preference to prevent startsWith error